[refactor] auth: 서비스 Command로 구분해서 리네이밍
